remove cfg from audio implementation & fix echo cancelling
